Bugs was fixed

+++ src/main/java/com/example/smarthome/server/exceptions/OutputNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class OutputNotFoundException extends Exception {
 
-    public OutputNotFoundException(int id) {
+    public OutputNotFoundException(long id) {
         super(String.format("Not found output with id '%d'", id));
     }
 }
